Minor change in Trajectories class
function weight now returns a int instead of double

+++ src/main/java/tt/euclidtime3i/trajectory/Trajectories.java
@@ -14,7 +14,7 @@ public class Trajectories {
     private Trajectories() {
     }
 
-    public static double weight(Collection<Straight> straights){
+    public static int weight(Collection<Straight> straights){
         int weight = 0;
         for (Straight straight : straights) {
             weight += straight.duration();
